refactor: split validation_test.go into per-resource test files
